Renamed package to lowercase
