+++ src/main/scala/org/orleans/developer/crypto/Hasher.scala
@@ -4,8 +4,8 @@ import java.security.MessageDigest
 
 import org.orleans.client.CryptoMain.getHash
 import org.orleans.developer.crypto.CryptoMessages.{FindMessage, MessageFound, MessageNotFound}
-import org.orleans.silo.Services.Grain.Grain
-import org.orleans.silo.Services.Grain.Grain.Receive
+import org.orleans.silo.services.grain.Grain
+import org.orleans.silo.services.grain.Grain.Receive
 import org.orleans.silo.dispatcher.Sender
 
 class Hasher(id: String) extends Grain(id){
