+++ src/main/scala/org/orleans/developer/twitter/Twitter.scala
@@ -1,7 +1,7 @@
 package org.orleans.developer.twitter
 import org.orleans.developer.twitter.TwitterMessages._
-import org.orleans.silo.Services.Grain.Grain
-import org.orleans.silo.Services.Grain.Grain.Receive
+import org.orleans.silo.services.grain.Grain
+import org.orleans.silo.services.grain.Grain.Receive
 import org.orleans.silo.dispatcher.Sender
 
 class Twitter(id: String) extends Grain(id) {
