+++ src/main/scala/org/orleans/developer/twitter/TwitterAcountRef.scala
@@ -1,7 +1,7 @@
 package org.orleans.developer.twitter
 import org.orleans.client.OrleansRuntime
 import org.orleans.developer.twitter.TwitterMessages._
-import org.orleans.silo.Services.Grain.{Grain, GrainRef, GrainReference}
+import org.orleans.silo.services.grain.{Grain, GrainRef, GrainReference}
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
