+++ src/main/scala/org/orleans/silo/Main.scala
@@ -2,8 +2,8 @@ package org.orleans.silo
 
 import ch.qos.logback.classic.Level
 import org.orleans.developer.twitter.{Twitter, TwitterAccount}
-import org.orleans.silo.Services.Grain.GrainRef
-import org.orleans.silo.Test.GreeterGrain
+import org.orleans.silo.services.grain.GrainRef
+import org.orleans.silo.test.GreeterGrain
 import org.orleans.silo.storage.GrainDatabase
 import org.orleans.silo.utils.ServerConfig
