+++ src/main/scala/org/orleans/silo/storage/MongoGrainDatabase.scala
@@ -5,7 +5,7 @@ import com.typesafe.scalalogging.LazyLogging
 import org.mongodb.scala._
 import org.mongodb.scala.model.Filters._
 import org.mongodb.scala.model.FindOneAndUpdateOptions
-import org.orleans.silo.Services.Grain.Grain
+import org.orleans.silo.services.grain.Grain
 import org.slf4j.LoggerFactory
 
 import scala.concurrent.ExecutionContext.Implicits.global
