+++ src/main/scala/org/orleans/silo/test/Test.scala
@@ -1,4 +1,4 @@
-package org.orleans.silo.Test
+package org.orleans.silo.test
 
 import java.util.concurrent.ConcurrentHashMap
